+++ src/components/CartPage/CartProductSummary/CartProductSummary.tsx
@@ -1,13 +1,13 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import { styled } from 'styled-components';
 import CartProductItem from '../CartProductItem/CartProductItem';
-import checkIcon from '../../assets/check.svg';
+import checkIcon from '../../../assets/check.svg';
 import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
 import { cartListAtom, itemSelectionAtom } from '../../../stores/cartItemsStore';
 import { useEffect, useState } from 'react';
 import useUpdateCartItems from '../../../hooks/useUpdateCartItems';
 import useGetData from '../../../hooks/useGetData';
 import { CartItem } from '../../../types';
+import * as Styled from './CartProductSummary.styles';
 
 const CartProductSummary = () => {
   const { data: cartListData, getData: getCartList } = useGetData<CartItem[]>('/cart-items');
@@ -66,11 +66,11 @@ const CartProductSummary = () => {
   return (
     <>
       {cartList.length !== 0 ? (
-        <Wrapper>
-          <ProductCountWrapper>
-            <ProductCount>든든배송 상품 ({cartList.length}개)</ProductCount>
-          </ProductCountWrapper>
-          <ProductList role='list'>
+        <Styled.Wrapper>
+          <Styled.ProductCountWrapper>
+            <Styled.ProductCount>든든배송 상품 ({cartList.length}개)</Styled.ProductCount>
+          </Styled.ProductCountWrapper>
+          <Styled.ProductList role='list'>
             {cartList.map((cartItem, index) => {
               return (
                 <CartProductItem
@@ -82,116 +82,27 @@ const CartProductSummary = () => {
                 />
               );
             })}
-          </ProductList>
-          <ProductSelectWrapper>
-            <Label>
-              <Input
+          </Styled.ProductList>
+          <Styled.ProductSelectWrapper>
+            <Styled.Label>
+              <Styled.Input
                 type='checkbox'
                 icon={checkIcon}
                 checked={itemSelection.every((item) => item === true)}
                 onChange={itemSelection.every((item) => item === true) ? cancelSelectAll : selectAll}
               />
-              <LabelContent>
+              <Styled.LabelContent>
                 전체선택 ({itemSelection.filter((item) => item === true).length}/{cartList.length})
-              </LabelContent>
-            </Label>
-            <ProductsDeleteButton onClick={deleteCartItem}>선택삭제</ProductsDeleteButton>
-          </ProductSelectWrapper>
-        </Wrapper>
+              </Styled.LabelContent>
+            </Styled.Label>
+            <Styled.ProductsDeleteButton onClick={deleteCartItem}>선택삭제</Styled.ProductsDeleteButton>
+          </Styled.ProductSelectWrapper>
+        </Styled.Wrapper>
       ) : (
-        <EmptyProductWrapper>장바구니가 비어있습니다!</EmptyProductWrapper>
+        <Styled.EmptyProductWrapper>장바구니가 비어있습니다!</Styled.EmptyProductWrapper>
       )}
     </>
   );
 };
-
-const Wrapper = styled.div`
-  width: 100%;
-`;
-
-const ProductCountWrapper = styled.div`
-  padding-bottom: 16px;
-  border-bottom: 4px solid #aaaaaa;
-`;
-
-const ProductCount = styled.span`
-  font-size: 20px;
-  line-height: 33px;
-  letter-spacing: 0.5px;
-  color: #333333;
-`;
-
-const ProductList = styled.ul`
-  display: flex;
-  flex-direction: column;
-`;
-
-const Label = styled.label`
-  width: fit-content;
-  display: flex;
-  align-items: center;
-`;
-
-const Input = styled.input<{ icon: string }>`
-  appearance: none;
-  -webkit-appearance: none;
-  width: 28px;
-  height: 28px;
-  border: 1px solid #22a6a2;
-  border-radius: 2px;
-  transform: translateY(0%);
-  cursor: pointer;
-
-  &:checked {
-    background: #333333;
-    border: 1px solid #3288ff;
-  }
-
-  &:after {
-    content: '';
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    left: 0;
-    top: 0;
-    background-image: ${({ icon }) => `url(${icon})`};
-    background-size: 20px;
-    background-repeat: no-repeat;
-    background-position: center;
-  }
-`;
-
-const LabelContent = styled.span`
-  margin-left: 15px;
-  font-size: 16px;
-  line-height: 20px;
-  letter-spacing: 0.5px;
-  color: #333333;
-  cursor: pointer;
-`;
-
-const ProductsDeleteButton = styled.button`
-  border: 1px solid #bbbbbb;
-  padding: 8px 18px;
-  background-color: #fff;
-  cursor: pointer;
-`;
-
-const ProductSelectWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 16px;
-`;
-
-const EmptyProductWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 20px;
-  line-height: 20px;
-  letter-spacing: 0.5px;
-  color: #333333;
-`;
 
 export default CartProductSummary;
